[Diagnostics] Removing Diagnostics property from `Parser` and `Binder` and have each `SyntaxNode` and `BoundNode` carry the diagnostics information instead (#36)
This change relieves the duty of `Parser` and `Binder` to hold the diagnostics information. These information will be attached directly to the `SyntaxNode` and `BoundNode` objects by implementing `IDiagnosable` interface. If a `SyntaxNode` and `BoundNode` object contains child `IDiagnosable` objects, the diagnostics information will be aggregated to the current object as well, e.g. Calling `GetDiagnostics()` on an `ExpressionStatement` object will get back all the diagnostics information from the inner expression as well.